requirements.txt
onnx
onnxscript
einops
torch_tensorrt
pillow
torchvision
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/
onnxruntime-gpu